메뉴 기능 보완

+++ MyMud/Player.cs
@@ -5,12 +5,12 @@ namespace MyMud
     internal class Player
     {
         
-        internal int playerMoney; // 플레이어의 돈
+        internal float playerMoney; // 플레이어의 돈
         int playerTotalMoney; // 플레이어의 코인 수익(코인 가격 * 수량) + 돈
         public List<Inventory> playerCoins = new List<Inventory>(); // 플레이어의 인벤토리
         
 
-        public Player(int v)
+        public Player(float v)
         {
             this.playerMoney = v;
         }
